Added JavaScript to draw table
The table is now drawn using JavaScript, which will prove itself
to be useful when the table needs to be redrawn whenever the user
updates it. Note that the table still exists in index.html only as a
fallback for users with JavaScript disabled. I plan on adding a message
in a <noscript> tag with some sample agendas.