Update license information
The license of AgendaCreator has been changed and is now more
well-documented. The new license of AgendaCreator is a dual-license between
Apache 2.0 or later and Mozilla Public License 2.0 or later, but all code that
was previously available under the MIT License continues to be available under
the MIT License until it is modified (in that case, older revisions will
continue to be available under the MIT License).