Restructure workflow and code structure
The biggest change in this commit is the addition of a package.json file for
npm and the restructure of the file hierarchy. These changes have been made so
that any needed JavaScript libraries can be added easily as dependencies and
updating the GitHub Pages site can now be automated. I have also added a
tsconfig.json file for TypeScript, and some serialization code using
FileSaver.js. Additionally, I have changed innerHTML to textContent in some
places in the code to prevent XSS attacks.